Add a ToSql method to BaseQuery to expose query args
The BaseQuery.String method returns the prepared SQL but does not return
the query arguments, making it hard to test queries without mocking the
database. By adding a BaseQuery.ToSql method that returns the args as
well as an error returned while compiling, anyone can easily test
queries built by a helper function and verify the sql and args are
correct.